UI fixes

+++ modules/nurses/delete.php
@@ -1,6 +1,11 @@
 <?php
+session_start();
+if (!isset($_SESSION['user_id'])) {
+    header("Location: ../../login.php");
+    exit();
+}
 include "../../config/db.php";
-$id = $_GET['id'];
+$id = intval($_GET['id']);
 $conn->query("DELETE FROM nurses WHERE nurse_id = $id");
 header("Location: view.php");
 exit();
